// apps/sync-service/src/common/types.ts
export type MappingItem =
  | "entities"
  | "user_changelogs"
  | "entity_material_activities"
  | "materials"
  | "batches"
  | "orders"
  | "order_items"
  | "order_comments"
  | "order_histories"
  | "order_item_projection_capacities"
  | "stocks"
  | "stock_exterminations"
  | "transactions"
  | "patients"
  | "users"
  | "manufactures"
  | "budget_sources"
  | "activities"
  | "transaction_reasons"
  | "programs"
  | "extermination_shipments"
  | "extermination_shipment_items"
  | "extermination_shipment_stocks"
  | "extermination_shipment_comments"
  | "extermination_transactions"
  | "stock_opnames"
  | "stock_opname_periods"
  | "reconciliations"
  | "assets"

export type MasterData =
  | "entities"
  | "materials"
  | "users"
  | "budget_sources"
  | "manufactures"